SeqGAN-Poem-torch-PARL/requirements.txt
paddlepaddle-gpu==1.8.2.post107
parl==1.3.1
torch==1.3.1
torchvision==0.4.2
